% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_phyto.R
\name{calc_phyto}
\alias{calc_phyto}
\title{Rough indication of required years of phyto-remediation}
\usage{
calc_phyto(dt, loc, pal, psd, qmax, pox, tpal, tpsd, extraction)
}
\arguments{
\item{dt}{data.table with soil data (topsoil only)}

\item{loc}{#column distinguishing locations}

\item{pal}{Column of P-AL (mg P2O5/100g)}

\item{psd}{Column of PSDox (%)}

\item{qmax}{Column of Qmax (mmol/kg)}

\item{pox}{Column of P-ox (mmol/kg)}

\item{tpal}{P-AL target (mg P2O5/100g)}

\item{tpsd}{PSDox target (%)}

\item{extraction}{#Assumed P extraction by the crop (kg P2O5/ha/year)}

\item{OS}{Column of organic matter (%)}

\item{thickness}{Column specificying thickness of soil layer (cm)}
}
\description{
Calculate require years of phyto-remediation based on initial P-AL (empirical, roeghoorn) and PSDox in combination with Qmax (theoretical)
}
\examples{
calc_phyto <- loadbasemap(dt = topsoildata, loc = "locatie", pal = "P-AL (mg P2O5/100g)", psd = "FVG (\%)", qmax = "Qmax (mmol/kg)",
pox = "P-ox (mmol/kg)",  OS = "Z_OM1_OF", thickness = "thick", tpal = 5, tpsd = 7, extraction = 70)

}
