% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_knmi.R
\name{dist_knmi}
\alias{dist_knmi}
\title{Retrieve information on the distance of KNMI weatherstations}
\usage{
dist_knmi(sf)
}
\arguments{
\item{sf}{shape (centroid is used)}
}
\description{
Retrieve information on the distance of KNMI weatherstations
}
\examples{
knmi <- dist_knmi(sf)
}
