% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadahn.R
\name{loadahn}
\alias{loadahn}
\title{Retrieve AHN height map}
\usage{
loadahn(sf, product, expand)
}
\arguments{
\item{sf}{Spatial feature object}

\item{product}{AHN product ("dtm_05m", "dsm_05m", "dtm_5m", "dsm_5m")}

\item{expand}{Extension in meters}
}
\description{
Retrieve AHN4 DEM for area of interest
}
\examples{
ahn <- loadahn(sf = parcel, expand = 20)

}
