% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadbasemap.R
\name{loadbasemap}
\alias{loadbasemap}
\title{Retrieve Aerial basemaps}
\usage{
loadbasemap(sf, source, expand)
}
\arguments{
\item{sf}{Spatial feature object}

\item{source}{either "google" or "esri" as a source of the basemap}

\item{expand}{value > 1 expands the basemap}
}
\description{
Load (and save to /dev) areal basemaps using the bbox of an sf object
}
\examples{
basemap <- loadbasemap(parcels, "google", 1.2)
}
