% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadbd50.R
\name{loadbd50}
\alias{loadbd50}
\title{Retrieve Alterra soil map}
\usage{
loadbd50(sf, expand, mask)
}
\arguments{
\item{sf}{Spatial feature polygon}

\item{expand}{Extension in meters}

\item{mask}{Boolian stating if desiring to mask the soil map to area of interest}
}
\description{
Retrieve 1:50.000 soil map for polygon of interest
}
\examples{
bd50 <- loadbd50(sf = parcel, expand = 20, mask = TRUE)
}
