% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadbodemschat.R
\name{loadbodemschat}
\alias{loadbodemschat}
\title{Retrieve Bodemschat}
\usage{
loadbodemschat(sf, expand)
}
\arguments{
\item{sf}{Spatial feature polygon}

\item{expand}{Extension in meters}
}
\description{
Retrieve Bodemschat for area of interest, rasterize to 25m spatial resolution
}
\examples{
bodemschat <- loadbodemschat(sf = aoi, expand = 0)

}
