% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadbrp.R
\name{loadbrp}
\alias{loadbrp}
\title{Retrieve BRP}
\usage{
loadbrp(sf, expand, year)
}
\arguments{
\item{sf}{Spatial feature object}

\item{expand}{Extension in meters. Defaults to 100 meters.}

\item{year}{2009 - 2026 supported. Defaults to 2019.}
}
\description{
Retrieve brp gewaspercelen for aoi using the bbox of an sf object
}
\examples{
brp <- loadbrp(sf = parcels, year = 2019, expand = 100)
}
