% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadgrondwatertrap.R
\name{loadgrondwatertrap}
\alias{loadgrondwatertrap}
\title{Retrieve grondwatertrappenkaart from bodemdata}
\usage{
loadgrondwatertrap(sf, year, res, expand, zoom)
}
\arguments{
\item{sf}{Spatial feature object}

\item{expand}{Extension in meters. Defaults to 20 meters.}

\item{zoom}{Zoom level. Defaults to 13}
}
\description{
Retrieve grondwatertrappenkaart using the bbox of an sf object
}
\examples{
basemap <- loadgrondwatertrap(sf = parcels, expand = 20, zoom = 16)
}
