% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadknmi.R
\name{loadknmi}
\alias{loadknmi}
\title{Retrieve daily KNMI data for a specific weather station}
\usage{
loadknmi(from_date, to_date, stationID, params)
}
\arguments{
\item{from_date}{Earliest date of the date range (yyyy.mm.dd)}

\item{to_date}{Latest date of the date range (yyyy.mm.dd)}

\item{stationID}{ID of the weather station}

\item{params}{Parameters of interest}
}
\description{
Retrieve daily KNMI data for a specific weather station
}
\examples{
prec_table <- loadknmi("2001.01.10", "2002.01.10", 240, c("RH", "RHX"))
}
