% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadlbk.R
\name{loadlbk}
\alias{loadlbk}
\title{Retrieve landschappelijke bodemkaart (LBK)}
\usage{
loadlbk(sf, expand)
}
\arguments{
\item{sf}{Spatial feature polygon}

\item{expand}{Extension in meters}
}
\description{
Retrieve landschappelijke bodemkaart for polygon of interest
}
\examples{
lbk <- loadlbk(sf = parcel, expand = 20)
}
