% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadox.R
\name{loadox}
\alias{loadox}
\title{Prep Eurofins oxalate data}
\usage{
loadox(file)
}
\arguments{
\item{file}{.csv file of Eurofins}
}
\description{
Retrieve p-ox, fe-ox and al-ox from eurofins .csv
}
\examples{
pox <- loadox("data/oxalate.csv")

}
