% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadwaterkwaliteit.R
\name{loadwaterkwaliteit}
\alias{loadwaterkwaliteit}
\title{Retrieve waterkwaliteitsdata}
\usage{
loadwaterkwaliteit(sf, expand)
}
\arguments{
\item{sf}{Spatial feature object}

\item{expand}{Extension in meters. Defaults to 20 meters.}
}
\description{
Retrieve waterkwaliteitsdata from waterkwaliteitsportaal (local dump)
}
\examples{
water <- loadwaterkwaliteit(sf = parcels, expand = 10000)
}
