% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_proj.R
\name{open_proj}
\alias{open_proj}
\title{Open project directory based on name}
\usage{
open_proj(name)
}
\arguments{
\item{name}{name of project as defined in 'projects' object}
}
\description{
Open project directory based on name
}
\examples{
open_proj(name = "hdsr automatisatie")

}
