% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_textuurdriehoek.R
\name{plot_textuurdriehoek}
\alias{plot_textuurdriehoek}
\title{Visualize texture triangles according to de Bakker & Schelling}
\usage{
plot_textuurdriehoek(textuurdata, type)
}
\arguments{
\item{textuurdata}{data.frame of texture data. Must contain columns 'clay', 'silt', 'sand'}

\item{type}{type of texture ternary triangle. Currently supported: 'eolisch'}
}
\description{
Create texture ternary plot in the ggplot environment
}
\examples{
plot_textuurdriehoek(textuurdata = textuur, type = "eolisch")
}
