% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_proj.R
\name{search_proj}
\alias{search_proj}
\title{Search projectfolders for text in order to find the right projectnumber}
\usage{
search_proj(range, text)
}
\arguments{
\item{range}{range in the projectnr according to structure in Projecten (full = all)}

\item{text}{text used in grepl, collapses with |}
}
\description{
Search projectfolders for text in order to find the right projectnumber
}
\examples{
search_proj(range = "1800-1900", text = c("borium", "mangaan"))

}
