% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_folderstructure.R
\name{set_folderstructure}
\alias{set_folderstructure}
\title{Set folder structure}
\usage{
set_folderstructure(folders)
}
\arguments{
\item{folders}{Vector of folders to create}
}
\description{
Create folders within working directory if not existing
}
\examples{
set_folderstructure(c("data", "dev", "output", "scripts"))

}
