% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadahn.R
\name{loadahn}
\alias{loadahn}
\title{Retrieve AHN height map}
\usage{
loadahn(sf, expand)
}
\arguments{
\item{sf}{Spatial feature object}

\item{expand}{Extension in meters}
}
\description{
Retrieve AHN4 50cm dtm height map for area of interest
}
\examples{
ahn <- loadahn(sf = parcel, expand = 20)

}
