% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadbasemap.R
\name{loadbasemap}
\alias{loadbasemap}
\title{Retrieve RGB basemaps from PDOK}
\usage{
loadbasemap(sf, year, res, expand, zoom)
}
\arguments{
\item{sf}{Spatial feature object}

\item{year}{2016 - 2023 supported. Defaults to most actual year.}

\item{res}{Spatial resolution (high or low). Defaults to high (8cm) for 2021-2023, low (25cm) for 2016-2021.}

\item{expand}{Extension in meters. Defaults to 20 meters.}

\item{zoom}{Zoom level. Defaults to 13}
}
\description{
Retrieve areal basemaps using the bbox of an sf object
}
\examples{
basemap <- loadbasemap(sf = parcels, year = 2023, res = 'high', expand = 20, zoom = 16)
}
