% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadbasemap.R
\name{loadbasemap}
\alias{loadbasemap}
\title{Retrieve Aerial basemaps}
\usage{
loadbasemap(sf, source, expand, zoom)
}
\arguments{
\item{sf}{Spatial feature object}

\item{source}{either "hybrid", "aerial", "topo" or "topo.v2 as a source of the basemap. Defaults to "hybrid".}

\item{expand}{Extension factor of the bounding box If 1, the bounding box is unchanged. Values smaller than 1 reduces the bounding box,
and values larger than 1 enlarges the bounding box. Defaults to 1.2.}

\item{zoom}{Zoom level (defaults to 18)}
}
\description{
Retrieve areal basemaps using the bbox of an sf object
}
\examples{
basemap <- loadbasemap(parcels, "google", 1.2, 18)

}
