% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadbasemap.R
\name{loadbasemap}
\alias{loadbasemap}
\title{Retrieve Aerial basemaps}
\usage{
loadbasemap(sf, source, expand)
}
\arguments{
\item{sf}{Spatial feature object}

\item{source}{either "google", "esri" or "osm" as a source of the basemap. Defaults to "google".}

\item{expand}{Extension factor of the bounding box If 1, the bounding box is unchanged. Values smaller than 1 reduces the bounding box,
and values larger than 1 enlarges the bounding box. Defaults to 1.2.}
}
\description{
Retrieve areal basemaps using the bbox of an sf object
}
\examples{
basemap <- loadbasemap(parcels, "google", 1.2)

}
